chronicle/tests/requirements.txt
pytest==7.2.2
pytest-django==3.9.0
pytest-cov==2.10.0

-r ../requirements.txt